#86 relocate and refactor helper function
