+++ applications/portal/frontend/src/components/Submit/CommercialForm.tsx
@@ -19,7 +19,7 @@ import { AlertIcon } from "../icons";
 import StepNavigation from "./StepNavigation";
 import { useFormik } from "formik";
 import * as yup from "yup";
-import { postNewAntibody } from "../../utils/antibody";
+import { postNewAntibody } from "../../helpers/antibody";
 
 const {
   bannerHeadingColor,
@@ -189,10 +189,11 @@ const FormLine = ({ children }) => {
 };
 const CommercialForm = (props) => {
   const classes = useStyles();
+  const { setAntibodyId, setApiResponse, next } = props;
 
   const postAntibody = (antibody) => {
     let ab = { ...antibody, type: "commercial" };
-    postNewAntibody(ab, props);
+    postNewAntibody(ab, setAntibodyId, setApiResponse, next);
   };
 
   const formik = useFormik({
@@ -508,7 +509,7 @@ const CommercialForm = (props) => {
       </Container>
       <StepNavigation
         previous={props.previous}
-        next={props.next}
+        next={next}
         hasPrevious={props.hasPrevious}
         isLastStep={true}
         activeStep={2}
