+++ applications/portal/frontend/src/components/Submit/PersonalForm.tsx
@@ -17,7 +17,7 @@ import { AlertIcon } from "../icons";
 import StepNavigation from "./StepNavigation";
 import { useFormik } from "formik";
 import * as yup from "yup";
-import { postNewAntibody } from "../../utils/antibody";
+import { postNewAntibody } from "../../helpers/antibody";
 
 const { bannerHeadingColor, primaryTextColor } = vars;
 
@@ -156,9 +156,11 @@ const FormLine = ({ children }) => {
 const PersonalForm = (props) => {
   const classes = useStyles();
 
+  const { setAntibodyId, setApiResponse, next } = props;
+
   const postAntibody = (antibody) => {
     let ab = { ...antibody, type: "personal" };
-    postNewAntibody(ab, props);
+    postNewAntibody(ab, setAntibodyId, setApiResponse, next);
   };
 
   const formik = useFormik({
@@ -404,7 +406,7 @@ const PersonalForm = (props) => {
       </Container>
       <StepNavigation
         previous={props.previous}
-        next={props.next}
+        next={next}
         hasPrevious={props.hasPrevious}
         isLastStep={true}
         activeStep={2}
